refactor: common 패키지 구조 변경
common이라는 상위 개념이 있기 때문에, 세부 단위 보다는 레이어로 나누는게 더 적절하다고 판단.

hiro032/issue-tracker#72
