+++ be/issue-tracker-be/src/main/java/com/codesquad/issuetracker/common/controller/CommonExceptionHandler.java
@@ -1,5 +1,6 @@
-package com.codesquad.issuetracker.common.exception;
+package com.codesquad.issuetracker.common.controller;
 
+import com.codesquad.issuetracker.common.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
